+++ firmware/boron-ins-fsm/src/imDoorSensor.cpp
@@ -70,9 +70,11 @@ doorData checkIM() {
         // Checks if the 0th bit (counting from 0) of doorStatus is 1
         doorTamperedFlag = (currentDoorData.doorStatus & 0b0001) != 0;
 
-        // Set timer starting when door closes
-        if ((previousDoorData.doorStatus & 0b0010) != 0 && (currentDoorData.doorStatus & 0b0010) == 0) {
-            timeWhenDoorClosed = millis();
+        // Reset timer on receiving a door close message or transition from open to closed + heartbeat
+        if ((currentDoorData.doorStatus & 0b0010) == 0) {
+            if ((currentDoorData.doorStatus & 0b1000) == 0 || (previousDoorData.doorStatus & 0b0010) != 0) {
+                timeWhenDoorClosed = millis();
+            }
         }
 
         // After a certain thereshold, the next door message received will trigger a boron heartbeat