+++ firmware/boron-ins-fsm/src/stateMachine.h
@@ -17,10 +17,10 @@
 // initial (default) values for state machine, can be changed via console function
 // or by writing something other than 0x8888 to the above flag in flash
 #define INS_THRESHOLD                   60
-#define STATE0_OCCUPANT_DETECTION_TIMER 3600000  // ms = 60 min
-#define STATE1_MAX_TIME                 15000    // ms = 15s
-#define STATE2_MAX_DURATION             1200000  // ms = 20 min
-#define STATE3_MAX_STILLNESS_TIME       120000   // ms = 2 minutes
+#define STATE0_OCCUPANT_DETECTION_TIMER 172800000  // 2 days
+#define STATE1_MAX_TIME                 15000      // ms = 15s
+#define STATE2_MAX_DURATION             1200000    // ms = 20 min
+#define STATE3_MAX_STILLNESS_TIME       120000     // ms = 2 minutes
 
 // How often to publish Heartbeat messages
 #define SM_HEARTBEAT_INTERVAL 660000  // ms = 11 min