+++ CHANGELOG.md
@@ -14,7 +14,7 @@ the code was deployed.
 ### Added
 
 - CORS configuration to Express Proxy Middleware.
-- Added a timer in state0 that is reset when the door closes where the sensor is allowed to detect for occupation (CU-8678t4ztg).
+- Added a timer in state0 that is reset when the door closes during which the sensor is allowed to detect for occupation (CU-8678t4ztg).
 - Added console function for the occupation detection timer as well as the timer count and timeout value in Debug Publishes (CU-8678t4ztg).
 - "doorMissedFrequently" field to JSON data submitted by brave sensors that is sent to /api/heartbeat (CU-860rk8v2a).
 - Sentry log in the case that doorMissedFrequently is true in posted data from brave sensor (CU-860rk8v2a).